+++ src/api/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to
 
 - Add tariffs support
 
+### Changed
+
+- Require at least one target when creating a tariff
+
 ## [0.34.1] - 2026-07-10
 
 ### Changed